+++ source/app/services/ExemploFuncional.js
@@ -1,5 +1,5 @@
 'use strict'
-
+//TODO: excluir
 const { Client } = require('@elastic/elasticsearch')
 const client = new Client({ node: 'http://localhost:9200' })
